patch: improve distance still Unclear as to the exact reason this seems faster without a disassembly of sqrt or sqrtf It appears however to be removing two sse cast instructions. see #259

+++ examples/custom_control/main.cpp
@@ -15,8 +15,8 @@ auto background = rbox(bkd_color, 5);
 // Function to calculate distance
 float distance(point a, point b)
 {
-  double c = (b.x - a.x) * (b.x - a.x) + (b.y - a.y) * (b.y - a.y);
-  return sqrt(c);
+  float c = (b.x - a.x) * (b.x - a.x) + (b.y - a.y) * (b.y - a.y);
+  return sqrtf(c);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
